porting changes made on stale branch to this one
