+++ src/react-boilerplate/webpack/webpack.production.config.js
@@ -1,7 +1,7 @@
 var webpack = require('webpack');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
 var baseConfig = require('./webpack.base.config.js');
-var StageDependancies = require('./stageDependancies');
+var DeployToPublicWebFolder = require('./DeployToPublicWebFolder');
 var buildPath = ['../laravel/public/build/', require('../package.json').name].join('');
 
 var config = {
@@ -22,7 +22,12 @@ var config = {
         warnings: true
       }
     }),
-    new StageDependancies( {appendTo:'.terminal', default: true, manifestFile: [buildPath, 'manifest.json'].join('/') })
+    new DeployToPublicWebFolder({
+      appendTo:'.terminal',
+      default: true,
+      manifestFile: [buildPath, 'manifest.json'].join('/'),
+      urlMapping: 'terminal'
+    })
   ]
 };
